whatpm/t/tree-test-1.dat

#data
<!DOCTYPE html><html title="t"><head><title></title></head><body></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   title="t"
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE html><html><head title="t"><title></title></head><body></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     title="t"
|     <title>
|   <body>

#data
<!DOCTYPE html><html><head><title class="t"></title></head><body></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|       class="t"
|   <body>

#data
<!DOCTYPE html><html><head></head><body><title class="a"></title></body></html>
#errors
title in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|       class="a"
|   <body>

#data
<!DOCTYPE html><html><head></head><body></body><title class="a"></title></html>
#errors
data after body
title in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|       class="a"
|   <body>

#data
<!DOCTYPE html><html><head><title></title></head><body title="t"></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     title="t"

#data
<!DOCTYPE html><html><head><title></title><meta name="a"></head><body></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|     <meta>
|       name="a"
|   <body>

#data
<!DOCTYPE html><html><head><title></title></head><body><meta name="a"></body></html>
#errors
meta in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|     <meta>
|       name="a"
|   <body>

#data
<!DOCTYPE html><html><head><title></title></head><body></body><meta name="a"></html>
#errors
data after body
meta in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|     <meta>
|       name="a"
|   <body>

#data
<!DOCTYPE html><html><head><title></title><link rel="a"></head><body></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|     <link>
|       rel="a"
|   <body>

#data
<!DOCTYPE html><html><head><title></title></head><body><link rel="a"></body></html>
#errors
link in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|     <link>
|       rel="a"
|   <body>

#data
<!DOCTYPE html><html><head><title></title></head><body></body><link rel="a"></html>
#errors
data after body
link in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|     <link>
|       rel="a"
|   <body>

#data
<!DOCTYPE html><html><head><title></title><base href="http://"></head><body></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|     <base>
|       href="http://"
|   <body>

#data
<!DOCTYPE html><html><head><title></title></head><body><base href="http://"></body></html>
#errors
base in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|     <base>
|       href="http://"
|   <body>

#data
<!DOCTYPE html><html><head><title></title></head><body></body><base href="http://"></html>
#errors
data after body
base in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|     <base>
|       href="http://"
|   <body>

#data
<!DOCTYPE html><p><b><p>X
#errors
<b> is not closed before second <p>
<b> is not closed before end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|       <b>
|     <p>
|       <b>
|         "X"

#data
<!DOCTYPE html><b><p>
#errors
<b> not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <b>
|       <p>

#data
<!DOCTYPE html><b><p>X
#errors
<b> not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <b>
|       <p>
|         "X"

#data
<link rel="a" href="b">
#errors
link in body
#document-fragment head
| <link>
|   href="b"
|   rel="a"

#data
<meta name="a" content="b">
#errors
meta in body
#document-fragment head
| <meta>
|   content="b"
|   name="a"

#data
<base href="a" target="b">
#errors
base in body
#document-fragment head
| <base>
|   href="a"
|   target="b"

#data
<base href="a" target="b">a
#errors
base in body
#document-fragment head
| <base>
|   href="a"
|   target="b"
| "a"
